+++ cmd/traffic/cmd/manager/internal/mutator/agent_injector_test.go
@@ -3,6 +3,7 @@ package mutator
 import (
 	"context"
 	"encoding/json"
+	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -22,6 +23,7 @@ import (
 const serviceAccountMountPath = "/var/run/secrets/kubernetes.io/serviceaccount"
 
 func TestTrafficAgentInjector(t *testing.T) {
+	type svcFinder func(c context.Context, client *kates.Client, portNameOrNumber, svcName, namespace string, labels map[string]string) (*kates.Service, error)
 	env := &managerutil.Env{
 		User:        "",
 		ServerHost:  "tel-example",
@@ -60,27 +62,42 @@ func TestTrafficAgentInjector(t *testing.T) {
 			},
 		},
 	}
-	numericPortSvc := &kates.Service{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Service",
-			APIVersion: "v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        "some-name",
-			Namespace:   "some-ns",
-			Labels:      nil,
-			Annotations: nil,
-		},
-		Spec: corev1.ServiceSpec{
-			Ports: []corev1.ServicePort{{
-				Protocol:   "TCP",
-				Port:       80,
-				TargetPort: intstr.FromInt(8888),
-			}},
-			Selector: map[string]string{
-				"service": "some-name",
+	defaultSvcFinder := func(c context.Context, client *kates.Client, portNameOrNumber, svcName, namespace string, labels map[string]string) (*kates.Service, error) {
+		return defaultSvc, nil
+	}
+	numericPortSvcFinder := func(c context.Context, client *kates.Client, portNameOrNumber, svcName, namespace string, labels map[string]string) (*kates.Service, error) {
+		return &kates.Service{
+			TypeMeta: metav1.TypeMeta{
+				Kind:       "Service",
+				APIVersion: "v1",
 			},
-		},
+			ObjectMeta: metav1.ObjectMeta{
+				Name:        "some-name",
+				Namespace:   "some-ns",
+				Labels:      nil,
+				Annotations: nil,
+			},
+			Spec: corev1.ServiceSpec{
+				Ports: []corev1.ServicePort{{
+					Protocol:   "TCP",
+					Port:       80,
+					TargetPort: intstr.FromInt(8888),
+				}},
+				Selector: map[string]string{
+					"service": "some-name",
+				},
+			},
+		}, nil
+	}
+	multiSvcFinder := func(c context.Context, client *kates.Client, portNameOrNumber, svcName, namespace string, labels map[string]string) (*kates.Service, error) {
+		// simulate not being given a service name and finding multiple services
+		if svcName == "" {
+			return nil, fmt.Errorf("multiple services found")
+		}
+		if svcName == defaultSvc.Name {
+			return defaultSvc, nil
+		}
+		return nil, fmt.Errorf("no services found")
 	}
 
 	tests := []struct {
@@ -88,7 +105,7 @@ func TestTrafficAgentInjector(t *testing.T) {
 		request       *admission.AdmissionRequest
 		expectedPatch string
 		expectedError string
-		service       *kates.Service
+		serviceFinder svcFinder
 	}{
 		{
 			"Skip Precondition: Not the right type of resource",
@@ -99,14 +116,14 @@ func TestTrafficAgentInjector(t *testing.T) {
 			}),
 			"",
 			"",
-			defaultSvc,
+			defaultSvcFinder,
 		},
 		{
 			"Error Precondition: Fail to unmarshall",
 			toAdmissionRequest(podResource, "I'm a string value, not an object"),
 			"",
 			"could not deserialize pod object",
-			defaultSvc,
+			defaultSvcFinder,
 		},
 		{
 			"Skip Precondition: No annotation",
@@ -115,7 +132,7 @@ func TestTrafficAgentInjector(t *testing.T) {
 			}),
 			"",
 			"",
-			defaultSvc,
+			defaultSvcFinder,
 		},
 		{
 			"Skip Precondition: No name/namespace",
@@ -126,7 +143,7 @@ func TestTrafficAgentInjector(t *testing.T) {
 			}),
 			"",
 			"",
-			defaultSvc,
+			defaultSvcFinder,
 		},
 		{
 			"Skip Precondition: Sidecar already injected",
@@ -157,10 +174,10 @@ func TestTrafficAgentInjector(t *testing.T) {
 			}),
 			"",
 			"",
-			defaultSvc,
+			defaultSvcFinder,
 		},
 		{
-			"Skip Precondition: No port specified",
+			"Error Precondition: No port specified",
 			toAdmissionRequest(podResource, corev1.Pod{
 				ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{
 					install.InjectAnnotation: "enabled",
@@ -172,11 +189,11 @@ func TestTrafficAgentInjector(t *testing.T) {
 				},
 			}),
 			"",
-			"",
-			defaultSvc,
+			"found no Service with a port that matches any container in this workload",
+			defaultSvcFinder,
 		},
 		{
-			"Skip Precondition: Sidecar has port collision",
+			"Error Precondition: Sidecar has port collision",
 			toAdmissionRequest(podResource, corev1.Pod{
 				ObjectMeta: metav1.ObjectMeta{
 					Annotations: map[string]string{
@@ -196,8 +213,8 @@ func TestTrafficAgentInjector(t *testing.T) {
 				},
 			}),
 			"",
-			"",
-			defaultSvc,
+			"is exposing the same port (9900) as the traffic-agent sidecar",
+			defaultSvcFinder,
 		},
 		{
 			"Apply Patch: Named port",
@@ -248,7 +265,112 @@ func TestTrafficAgentInjector(t *testing.T) {
 				`}}` +
 				`]`,
 			"",
-			defaultSvc,
+			defaultSvcFinder,
+		},
+		{
+			"Error Precondition: Multiple services",
+			toAdmissionRequest(podResource, corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Annotations: map[string]string{
+						install.InjectAnnotation: "enabled",
+					},
+					Labels: map[string]string{
+						"service": "some-name",
+					},
+					Namespace: "some-ns",
+					Name:      "some-name"},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{{
+						Name:  "some-app-name",
+						Image: "some-app-image",
+						Ports: []corev1.ContainerPort{{
+							Name: "http", ContainerPort: 8888},
+						}},
+					},
+				},
+			}),
+			"",
+			"multiple services found",
+			multiSvcFinder,
+		},
+		{
+			"Error Precondition: Invalid service name",
+			toAdmissionRequest(podResource, corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Annotations: map[string]string{
+						install.InjectAnnotation:      "enabled",
+						install.ServiceNameAnnotation: "khruangbin",
+					},
+					Labels: map[string]string{
+						"service": "some-name",
+					},
+					Namespace: "some-ns",
+					Name:      "some-name"},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{{
+						Name:  "some-app-name",
+						Image: "some-app-image",
+						Ports: []corev1.ContainerPort{{
+							Name: "http", ContainerPort: 8888},
+						}},
+					},
+				},
+			}),
+			"",
+			"no services found",
+			multiSvcFinder,
+		},
+		{
+			"Apply Patch: Multiple services",
+			toAdmissionRequest(podResource, corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Annotations: map[string]string{
+						install.InjectAnnotation:      "enabled",
+						install.ServiceNameAnnotation: defaultSvc.Name,
+					},
+					Labels: map[string]string{
+						"service": "some-name",
+					},
+					Namespace: "some-ns",
+					Name:      "some-name"},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{{
+						Name:  "some-app-name",
+						Image: "some-app-image",
+						Ports: []corev1.ContainerPort{{
+							Name: "http", ContainerPort: 8888},
+						}},
+					},
+				},
+			}),
+			`[` +
+				`{"op":"replace","path":"/spec/containers/0/ports/0/name","value":"tm-http"},` +
+				`{"op":"add","path":"/spec/containers/-","value":{` +
+				`"name":"traffic-agent",` +
+				`"image":"docker.io/datawire/tel2:2.3.1",` +
+				`"args":["agent"],` +
+				`"ports":[{"name":"http","containerPort":9900,"protocol":"TCP"}],` +
+				`"env":[` +
+				`{"name":"TELEPRESENCE_CONTAINER","value":"some-app-name"},` +
+				`{"name":"_TEL_AGENT_LOG_LEVEL","value":"info"},` +
+				`{"name":"_TEL_AGENT_NAME","value":"some-name"},` +
+				`{"name":"_TEL_AGENT_NAMESPACE","valueFrom":{"fieldRef":{"fieldPath":"metadata.namespace"}}},` +
+				`{"name":"_TEL_AGENT_POD_IP","valueFrom":{"fieldRef":{"fieldPath":"status.podIP"}}},` +
+				`{"name":"_TEL_AGENT_APP_PORT","value":"8888"},` +
+				`{"name":"_TEL_AGENT_PORT","value":"9900"},` +
+				`{"name":"_TEL_AGENT_MANAGER_HOST","value":"traffic-manager.default"}` +
+				`],` +
+				`"resources":{},` +
+				`"volumeMounts":[{"name":"traffic-annotations","mountPath":"/tel_pod_info"}],` +
+				`"readinessProbe":{"exec":{"command":["/bin/stat","/tmp/agent/ready"]}}` +
+				`}},` +
+				`{"op":"add","path":"/spec/volumes/-","value":{` +
+				`"name":"traffic-annotations",` +
+				`"downwardAPI":{"items":[{"path":"annotations","fieldRef":{"fieldPath":"metadata.annotations"}}]}` +
+				`}}` +
+				`]`,
+			"",
+			multiSvcFinder,
 		},
 		{
 			"Apply Patch: Numeric port",
@@ -310,7 +432,7 @@ func TestTrafficAgentInjector(t *testing.T) {
 				`}}` +
 				`]`,
 			"",
-			numericPortSvc,
+			numericPortSvcFinder,
 		},
 		{
 			"Apply Patch: Numeric port with init containers",
@@ -375,7 +497,7 @@ func TestTrafficAgentInjector(t *testing.T) {
 				`}}` +
 				`]`,
 			"",
-			numericPortSvc,
+			numericPortSvcFinder,
 		},
 		{
 			"Apply Patch: Numeric port re-processing",
@@ -431,7 +553,7 @@ func TestTrafficAgentInjector(t *testing.T) {
 				`}}` +
 				`]`,
 			"",
-			numericPortSvc,
+			numericPortSvcFinder,
 		},
 		{
 			"Apply Patch: volumes are copied",
@@ -489,7 +611,7 @@ func TestTrafficAgentInjector(t *testing.T) {
 				`}}` +
 				`]`,
 			"",
-			defaultSvc,
+			defaultSvcFinder,
 		},
 	}
 
@@ -500,12 +622,10 @@ func TestTrafficAgentInjector(t *testing.T) {
 			defer func() {
 				findMatchingService = fms
 			}()
-			findMatchingService = func(c context.Context, client *kates.Client, portNameOrNumber, svcName, namespace string, labels map[string]string) (*kates.Service, error) {
-				return test.service, nil
-			}
+			findMatchingService = test.serviceFinder
 
 			actualPatch, actualErr := agentInjector(ctx, test.request)
-			assertContains(t, actualErr, test.expectedError)
+			requireContains(t, actualErr, test.expectedError)
 			if actualPatch != nil || test.expectedPatch != "" {
 				patchBytes, err := json.Marshal(actualPatch)
 				require.NoError(t, err)
@@ -516,16 +636,16 @@ func TestTrafficAgentInjector(t *testing.T) {
 	}
 }
 
-func assertContains(t *testing.T, err error, expected string) {
+func requireContains(t *testing.T, err error, expected string) {
 	if expected == "" {
-		assert.NoError(t, err)
+		require.NoError(t, err)
 		return
 	}
 	if err == nil {
-		assert.Emptyf(t, expected, "expected error %q", expected)
+		require.Emptyf(t, expected, "expected error %q", expected)
 		return
 	}
-	assert.Contains(t, err.Error(), expected)
+	require.Contains(t, err.Error(), expected)
 }
 
 func toAdmissionRequest(resource metav1.GroupVersionResource, object interface{}) *admission.AdmissionRequest {